Refactor: response 객체들 디렉토리로 이전
