+++ src/dto/response/study-apply.response.ts
