+++ src/dto/response/study-team-member.response.ts
